+++ test/fixture/define/exports-multiple/output.js
@@ -1,2 +1,2 @@
 var a = 'baz';
-export { a as foo, a as bar }
+export { a as foo, a as bar };