+++ test/fixture/define/exports-same/output.js
@@ -1,3 +1,8 @@
-import bar from 'bar'
-import baz from 'baz'
-export { bar, baz }
+import bar from 'bar';
+import baz from 'baz';
+var a = bar;
+var b = baz;
+export {
+    a as bar,
+    b as baz
+};